Split comparisons into their own function from smart playlists
